pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize>=7.40.1",
    "autogen-agentchat>=0.5.6",
    "autogen-ext[openai]>=0.5.6",
    "black>=25.1.0",
    "chromadb>=1.0.10",
    "jsonschema>=4.17.3",
    "nltk>=3.9.1",
    "openinference-instrumentation-autogen>=0.1.9",
    "openinference-instrumentation-openai>=0.1.28",
    "prompt-toolkit>=3.0.51",
    "pylint>=3.3.7",
    "pyproject-toml>=0.1.0",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.11",
    "scikit-learn>=1.6.1",
]
[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
[tool.ruff.lint]
select = ["E", "F"]  # Enables Error and Pyflakes rules
# ignore = ["F401", "F841", "E501"]
